mixer: fix mixer_ch_get_pos overflow after 23 seconds
We forgot changing this when we switched to 64-bit positions
in bb2eff0f.

+++ src/audio/mixer.c
@@ -433,7 +433,7 @@ void mixer_ch_set_pos(int ch, float pos) {
 float mixer_ch_get_pos(int ch) {
 	mixer_channel_t *c = &Mixer.channels[ch];
 	assertf(!(c->flags & CH_FLAGS_STEREO_SUB), "mixer_ch_get_pos: cannot call on secondary stereo channel %d", ch);
-	uint32_t pos = c->pos >> (c->flags & CH_FLAGS_BPS_SHIFT);
+	uint64_t pos = c->pos >> (c->flags & CH_FLAGS_BPS_SHIFT);
 	return (float)pos / (float)(1<<MIXER_FX64_FRAC);
 }
